GH-53: Merge mask generator script into pattern generator
